+++ handler/comment/post_comment.go
@@ -12,13 +12,6 @@ import (
 	"k8s.io/klog"
 )
 
-// ListResponse comment list request
-type ListResponse struct {
-	StatusCode  int32                         `json:"status_code"`
-	StatusMsg   string                        `json:"status_msg,omitempty"`
-	CommentList []comment_service.CommentInfo `json:"comment_list,omitempty"`
-}
-
 // ActionResponse comment action response
 type ActionResponse struct {
 	StatusCode int32                       `json:"status_code"`
@@ -110,37 +103,4 @@ func Action(c *gin.Context) {
 		})
 		klog.Info("CommentController-Comment_Action: return delete success") //删除评论成功，返回正确信息
 	}
-}
-
-// List comment list
-// @Router /douyin/comment/list/ [get]
-func List(c *gin.Context) {
-	klog.Info("comment list")
-	id := c.Query("video_id")
-	klog.Info(id)
-	videoId, err := strconv.ParseInt(id, 10, 64)
-	if err != nil {
-		klog.Errorf("strconv.ParseInt error: %s", err)
-		c.JSON(http.StatusOK, ListResponse{
-			StatusCode: -1,
-			StatusMsg:  "comment videoId json invalid",
-		})
-		return
-	}
-	service := new(comment_service.CommentService)
-	commentList, err := service.GetList(int64(int(videoId)))
-	if err != nil {
-		c.JSON(http.StatusOK, ListResponse{
-			StatusCode: -1,
-			StatusMsg:  "comment list failed",
-		})
-		log.Println("CommentController-Comment_List: return comment list failed") //获取评论列表失败
-		return
-	}
-	c.JSON(http.StatusOK, ListResponse{
-		StatusCode:  0,
-		StatusMsg:   "comment list success",
-		CommentList: commentList,
-	})
-	klog.Info("CommentController-Comment_List: return comment list success")
 }
