+++ handler/follow/query_follow_list_test.go
@@ -1,7 +1,6 @@
-package test
+package follow
 
 import (
-	"douyin-lite/handler/follow"
 	"douyin-lite/repository"
 	"fmt"
 	"testing"
@@ -9,7 +8,7 @@ import (
 
 func TestQueryFollowListHandler(t *testing.T) {
 	repository.Init()
-	followListResp, err := follow.QueryFollowListHandler("2")
+	followListResp, err := QueryFollowListHandler("2")
 	if err != nil {
 		panic(err)
 	}
