+++ handler/user/register_user_test.go
@@ -1,7 +1,6 @@
-package test
+package user
 
 import (
-	"douyin-lite/handler/user"
 	"douyin-lite/repository"
 	"fmt"
 	"testing"
@@ -9,7 +8,7 @@ import (
 
 func TestRegisterUserHandler(t *testing.T) {
 	repository.Init()
-	registerResp, err := user.RegisterUserHandler("mao123", "123")
+	registerResp, err := RegisterUserHandler("mao123", "123")
 	if err != nil {
 		panic(err)
 	}
