+++ repository/following_test.go
@@ -1,14 +1,13 @@
-package test
+package repository
 
 import (
-	"douyin-lite/repository"
 	"fmt"
 	"testing"
 )
 
 func TestFollowingDao_CreateFollowing(t *testing.T) {
-	repository.Init()
-	followingDao := repository.NewFollowingDaoInstance()
+	Init()
+	followingDao := NewFollowingDaoInstance()
 	err := followingDao.CreateFollowing(1, 2)
 	if err != nil {
 		panic(err)
@@ -20,8 +19,8 @@ func TestFollowingDao_CreateFollowing(t *testing.T) {
 }
 
 func TestFollowingDao_DeleteFollowing(t *testing.T) {
-	repository.Init()
-	followingDao := repository.NewFollowingDaoInstance()
+	Init()
+	followingDao := NewFollowingDaoInstance()
 	err := followingDao.DeleteFollowing(1, 2)
 	if err != nil {
 		panic(err)
@@ -33,8 +32,8 @@ func TestFollowingDao_DeleteFollowing(t *testing.T) {
 }
 
 func TestFollowingDao_QueryFollowingListByHostId(t *testing.T) {
-	repository.Init()
-	followingDao := repository.NewFollowingDaoInstance()
+	Init()
+	followingDao := NewFollowingDaoInstance()
 	UserList, err := followingDao.QueryFollowingListByHostId(1)
 	if err != nil {
 		panic(err)
@@ -45,8 +44,8 @@ func TestFollowingDao_QueryFollowingListByHostId(t *testing.T) {
 }
 
 func TestFollowingDao_IncFollowingCnt(t *testing.T) {
-	repository.Init()
-	followingDao := repository.NewFollowingDaoInstance()
+	Init()
+	followingDao := NewFollowingDaoInstance()
 	err := followingDao.IncFollowingCnt(4)
 	if err != nil {
 		panic(err)
@@ -54,8 +53,8 @@ func TestFollowingDao_IncFollowingCnt(t *testing.T) {
 }
 
 func TestFollowingDao_DecFollowingCnt(t *testing.T) {
-	repository.Init()
-	followingDao := repository.NewFollowingDaoInstance()
+	Init()
+	followingDao := NewFollowingDaoInstance()
 	err := followingDao.DecFollowingCnt(1)
 	if err != nil {
 		panic(err)
@@ -63,8 +62,8 @@ func TestFollowingDao_DecFollowingCnt(t *testing.T) {
 }
 
 func TestFollowingDao_FollowAction(t *testing.T) {
-	repository.Init()
-	followingDao := repository.NewFollowingDaoInstance()
+	Init()
+	followingDao := NewFollowingDaoInstance()
 	err := followingDao.FollowAction(1, 2)
 	if err != nil {
 		panic(err)
@@ -84,8 +83,8 @@ func TestFollowingDao_FollowAction(t *testing.T) {
 }
 
 func TestFollowingDao_UnfollowAction(t *testing.T) {
-	repository.Init()
-	followingDao := repository.NewFollowingDaoInstance()
+	Init()
+	followingDao := NewFollowingDaoInstance()
 	err := followingDao.UnfollowAction(1, 2)
 	if err != nil {
 		panic(err)
