+++ repository/user_test.go
@@ -1,14 +1,13 @@
-package test
+package repository
 
 import (
-	"douyin-lite/repository"
 	"fmt"
 	"testing"
 )
 
 func TestUserDao_CreateUser(t *testing.T) {
-	repository.Init()
-	userDao := repository.NewUserDaoInstance()
+	Init()
+	userDao := NewUserDaoInstance()
 	userDao.CreateUser("wdp", 0, 0)
 	userDao.CreateUser("lwh", 0, 0)
 	userDao.CreateUser("cwp", 0, 0)
@@ -19,15 +18,15 @@ func TestUserDao_CreateUser(t *testing.T) {
 }
 
 func TestUserDao_CreateRegisterUser(t *testing.T) {
-	repository.Init()
-	userDao := repository.NewUserDaoInstance()
+	Init()
+	userDao := NewUserDaoInstance()
 	userDao.CreateRegisterUser("wdp", "123")
 	userDao.CreateRegisterUser("lwh", "456")
 }
 
 func TestUserDao_QueryIsUserExist(t *testing.T) {
-	repository.Init()
-	userDao := repository.NewUserDaoInstance()
+	Init()
+	userDao := NewUserDaoInstance()
 	isExist, err := userDao.QueryIsUserExist("wangdongdong")
 	if err != nil {
 		if isExist == false {
@@ -41,8 +40,8 @@ func TestUserDao_QueryIsUserExist(t *testing.T) {
 }
 
 func TestUserDao_QueryIsUserLogin(t *testing.T) {
-	repository.Init()
-	userDao := repository.NewUserDaoInstance()
+	Init()
+	userDao := NewUserDaoInstance()
 	qUser, err := userDao.QueryLoginUser("mao122", "123")
 	if err != nil {
 		panic(err)
@@ -51,8 +50,8 @@ func TestUserDao_QueryIsUserLogin(t *testing.T) {
 }
 
 func TestUserDao_QueryUserById(t *testing.T) {
-	repository.Init()
-	userDao := repository.NewUserDaoInstance()
+	Init()
+	userDao := NewUserDaoInstance()
 	qUser, err := userDao.QueryUserById(21)
 	if err != nil {
 		panic(err)
